Move WorkstationPageForm to WorkstationForm directory as a subcategory of WorkstationForm
